+++ src/index.ts
@@ -36,7 +36,7 @@ import * as configuration from "./configuration";
 import { getDomainMetrics } from "./actions/getDomainMetrics";
 import { getRegistrarForDomain } from "./helpers";
 import { Bid } from "./zAuction";
-import { ContentModerationResponse } from "./api/types";
+import { ContentModerationResponse } from "./types";
 
 export * from "./types";
 export { configuration };
